Update tech_test_1.py to match changes in the vector module
the script was unable to run due to changing math.floor() for flooring vectors to a an internal vec.floor() function, which caused the math.floor() functions to attempt to treat the vectors as a number, which caused an error. the (old) implementation of the vector module has been removed as well.